move getMiddleware
